[PM-39851] test: Fix FillAssistRepositoryTests for cache-readable sync-skip behavior
syncRules_withinUpdateInterval now also stubs cached data to match the
updated skip condition, and a new test covers a fresh timestamp with
no readable cache. Fixture builders moved to a separate file to stay
under the file length limit.
